tasks/ukri_resources/requirements.txt
click==8.0.3
CouchDB==1.2
psycopg2-binary==2.9.1
